// frontend/src/generated/api/models/ExperimentExportFormat.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EnergyForecast API
 * Energy consumption analysis and forecasting API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Bounded export representations for one experiment.
 * @export
 */
export const ExperimentExportFormat = {
    MetricsCsv: 'metrics_csv',
    MetricsJson: 'metrics_json',
    ManifestJson: 'manifest_json'
} as const;
export type ExperimentExportFormat = typeof ExperimentExportFormat[keyof typeof ExperimentExportFormat];


export function instanceOfExperimentExportFormat(value: any): boolean {
    for (const key in ExperimentExportFormat) {
        if (Object.prototype.hasOwnProperty.call(ExperimentExportFormat, key)) {
            if (ExperimentExportFormat[key as keyof typeof ExperimentExportFormat] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ExperimentExportFormatFromJSON(json: any): ExperimentExportFormat {
    return ExperimentExportFormatFromJSONTyped(json, false);
}

export function ExperimentExportFormatFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExperimentExportFormat {
    return json as ExperimentExportFormat;
}

export function ExperimentExportFormatToJSON(value?: ExperimentExportFormat | null): any {
    return value as any;
}

export function ExperimentExportFormatToJSONTyped(value: any, ignoreDiscriminator: boolean): ExperimentExportFormat {
    return value as ExperimentExportFormat;
}